ActivityPanel gefixt

+++ Wielwijk/src/wielwijk/ActivityPanel.java
@@ -60,7 +60,7 @@ public class ActivityPanel extends JPanel {
                 (Integer) map.get("upper_user_limit"),
                 (String) map.get("lld"),
                 (Boolean) map.get("cancelled"),
-                (Long) map.get("id")
+                ((Long) map.get("id")).intValue()
             );
             
             data.add(act);
